requirements.txt
django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-environ==0.12.0
django-filter==24.1
pillow==10.3.0
markdown==3.6.0
pyjwt==2.8.0
celery==5.4.0
gunicorn==21.2.0
whitenoise==6.7.0
drf-yasg==1.21.7
pytest-django==4.8.0
pytest==8.2.1
factory-boy==3.3.0
pytest-cov==4.1.0